Feature/improve (#1)
Optimize memory / object creation. No need to create three objects on context when we know we are going to need them for runtime, merge result / execution state into context.
Various performance optimizations.
Allow execution of static methods.
Allow interceptors to decide when we create the activations if they should run or not for a particular method.